giving 60 second leway for token refresh on frontend. organizing code
