+++ opencti-platform/opencti-front/src/private/Index.js
@@ -72,7 +72,7 @@ const Index = (me) => {
       });
     }
     const jwtToken = JSON.parse(atob(me.me.access_token.split('.')[1]));
-    const expiration = (jwtToken.exp * 1000) - Date.now();
+    const expiration = ((jwtToken.exp - 60) * 1000) - Date.now();
     if (expiration >= 0) {
       setInterval(() => {
         localStorage.removeItem('token');
